modul3_Python/12_Полезные_вещи_в_разработке/best_practices/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==0.21.1
pyyaml==6.0
pathlib==1.0.1